// src/enums/ValidationMessages.ts
/* eslint-disable @typescript-eslint/no-duplicate-enum-values */
export enum ValidationMessages {
  Accepted = "The :attribute must be accepted.",
  Array = "The :attribute must be an array.",
  BetweenArray = "The :attribute must have between :min and :max items.",
  BetweenFile = "The :attribute must be between :min and :max kilobytes.",
  BetweenNumeric = "The :attribute must be between :min and :max.",
  BetweenString = "The :attribute must be between :min and :max characters.",
  Boolean = "The :attribute field must be true or false.",
  Confirmed = "The :attribute confirmation does not match.",
  Decimal = "The :attribute must have :decimal decimal places.",
  Declined = "The :attribute must be declined.",
  Different = "The :attribute and :other must be different.",
  Digits = "The :attribute must be :digits digits.",
  DigitsBetween = "The :attribute must be between :min and :max digits.",
  DoesntEndWith = "The :attribute may not end with one of the following: :values.",
  DoesntStartWith = "The :attribute may not start with one of the following: :values.",
  Email = "The :attribute must be a valid email address.",
  EndsWith = "The :attribute must end with one of the following: :values.",
  File = "The :attribute must be a file.",
  In = "The selected :attribute is invalid.",
  InArray = "The :attribute field does not exist in :other.",
  Integer = "The :attribute must be an integer.",
  Lowercase = "The :attribute must be lowercase.",
  MaxArray = "The :attribute must not have more than :max items.",
  MaxFile = "The :attribute must not be greater than :max kilobytes.",
  MaxNumeric = "The :attribute must not be greater than :max.",
  MaxString = "The :attribute must not be greater than :max characters.",
  MaxDigits = "The :attribute must not have more than :max digits.",
  MinArray = "The :attribute must have at least :min items.",
  MinFile = "The :attribute must be at least :min kilobytes.",
  MinNumeric = "The :attribute must be at least :min.",
  MinString = "The :attribute must be at least :min characters.",
  MinDigits = "The :attribute must have at least :min digits.",
  MultipleOf = "The :attribute must be a multiple of :value.",
  NotIn = "The selected :attribute is invalid.",
  NotRegex = "The :attribute format is invalid.",
  Numeric = "The :attribute must be a number.",
  Regex = "The :attribute format is invalid.",
  Required = "The :attribute field is required.",
  RequiredIf = "The :attribute field is required when :other is :value.",
  RequiredUnless = "The :attribute field is required unless :other is in :values.",
  RequiredWith = "The :attribute field is required when :values is present.",
  RequiredWithAll = "The :attribute field is required when :values are present.",
  RequiredWithout = "The :attribute field is required when :values is not present.",
  RequiredWithoutAll = "The :attribute field is required when none of :values are present.",
  Same = "The :attribute and :other must match.",
  StartsWith = "The :attribute must start with one of the following: :values.",
  String = "The :attribute must be a string.",
  Uppercase = "The :attribute must be uppercase.",
  Ulid = "The :attribute must be a valid ULID.",
  Uuid = "The :attribute must be a valid UUID.",
}